+++ lib/fluid_dialog.dart
@@ -4,6 +4,8 @@
 /// - [FluidDialog] for creating a dialog.
 /// - [DialogNavigator] for more information about navigating a dialog.
 /// - [FluidDialogPage] for creating pages for a dialog.
+// ignore_for_file: comment_references
+
 library fluid_dialog;
 
 export 'src/dialog.dart';