+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## 1.0.0
+# 1.1.0
+
+* **Beaking**: Renamed `dialogDecoration` to `defaultDecoration`.
+* **Feat**: Added the ability to change the decoration of the dialog between routes.
+* **Fix**: Fixed transition default animation not working properly.
+* **Docs**: Improved docs and readme.
+* **Refactor**: Now using [very_good_analysis](https://pub.dev/packages/very_good_analysis).
+
+# 1.0.0
 
 * initial release.